+++ tools/test-js-with-node
@@ -73,6 +73,7 @@ EXEMPT_FILES = make_set(
         "web/src/compose_closed_ui.ts",
         "web/src/compose_fade.ts",
         "web/src/compose_notifications.ts",
+        "web/src/compose_paste.ts",
         "web/src/compose_recipient.ts",
         "web/src/compose_reply.ts",
         "web/src/compose_send_menu_popover.js",
@@ -86,7 +87,7 @@ EXEMPT_FILES = make_set(
         "web/src/condense.ts",
         "web/src/confirm_dialog.ts",
         "web/src/copied_tooltip.ts",
-        "web/src/copy_and_paste.ts",
+        "web/src/copy_messages.ts",
         "web/src/csrf.ts",
         "web/src/css_variables.ts",
         "web/src/custom_profile_fields_ui.ts",
