+++ web/src/hotkey.js
@@ -15,7 +15,7 @@ import * as compose_send_menu_popover from "./compose_send_menu_popover.js";
 import * as compose_state from "./compose_state.ts";
 import * as compose_textarea from "./compose_textarea.ts";
 import * as condense from "./condense.ts";
-import * as copy_and_paste from "./copy_and_paste.ts";
+import * as copy_messages from "./copy_messages.ts";
 import * as deprecated_feature_notice from "./deprecated_feature_notice.ts";
 import * as drafts_overlay_ui from "./drafts_overlay_ui.ts";
 import * as emoji from "./emoji.ts";
@@ -1144,7 +1144,7 @@ export function process_hotkey(e, hotkey) {
             navigate.page_down();
             return true;
         case "copy_with_c":
-            copy_and_paste.copy_handler();
+            copy_messages.copy_handler();
             return true;
     }
